+++ src/plone.server/plone/server/languages.py
@@ -1,8 +1,7 @@
 # -*- encoding: utf-8 -*-
+from plone.server import configure
 from plone.server.interfaces import IDefaultLayer
 from plone.server.interfaces import ILanguage
-from zope.component import adapter
-from zope.interface import implementer
 
 
 class IEN(ILanguage):
@@ -26,29 +25,33 @@ class DefaultLanguage(object):
         return self.context
 
 
-@adapter(IDefaultLayer)
-@implementer(IEN)
+@configure.adapter(
+    for_=IDefaultLayer,
+    provides=IEN)
 class EN(DefaultLanguage):
     def __init__(self, content):
         self.content = content
 
 
-@adapter(IDefaultLayer)
-@implementer(ICA)
+@configure.adapter(
+    for_=IDefaultLayer,
+    provides=ICA)
 class CA(DefaultLanguage):
     def __init__(self, content):
         self.content = content
 
 
-@adapter(IDefaultLayer)
-@implementer(IFI)
+@configure.adapter(
+    for_=IDefaultLayer,
+    provides=IFI)
 class FI(DefaultLanguage):
     def __init__(self, content):
         self.content = content
 
 
-@adapter(IDefaultLayer)
-@implementer(IENUS)
+@configure.adapter(
+    for_=IDefaultLayer,
+    provides=IENUS)
 class ENUS(DefaultLanguage):
     def __init__(self, content):
         self.content = content